Update README.md
In this project i created a page with HTML and styled the page with CSS.
I made page functions using Vanilla JS . 
I used color sections as switching buttons. 
When user click on any color page background and all the texts take on the same color.
This is my first Vanilla JS project. It was fun to use Vanilla.